+++ src/components/formcomponents/date/date.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import { parseISO } from 'date-fns';
 import { QuestionnaireItem, QuestionnaireResponseItemAnswer, Questionnaire } from 'fhir/r4';
-import { connect } from 'react-redux';
 import { ThunkDispatch } from 'redux-thunk';
 
 import { LanguageLocales } from '@helsenorge/core-utils/constants/languages';
@@ -18,18 +17,17 @@ import { safeParseJSON } from '../../../util/date-fns-utils';
 import { getExtension, getItemControlExtensionValue } from '../../../util/extension';
 import { evaluateFhirpathExpressionToGetDate } from '../../../util/fhirpathHelper';
 import { getLabelText, getSublabelText } from '../../../util/index';
-import { mapStateToProps, mergeProps, mapDispatchToProps } from '../../../util/map-props';
 import { Path } from '../../../util/refero-core';
 import { Resources } from '../../../util/resources';
 import ReactHookFormHoc, { FormProps } from '../../../validation/ReactHookFormHoc';
 import withCommonFunctions, { WithCommonFunctionsAndEnhancedProps } from '../../with-common-functions';
+import { useDispatch } from 'react-redux';
 
 export interface DateProps extends WithCommonFunctionsAndEnhancedProps, FormProps {
   item: QuestionnaireItem;
   questionnaire?: Questionnaire;
   answer: QuestionnaireResponseItemAnswer;
   resources?: Resources;
-  dispatch?: ThunkDispatch<GlobalState, void, NewValueAction>;
   path: Array<Path>;
   pdf?: boolean;
   language?: string;
@@ -48,7 +46,6 @@ const DateComponent = (props: React.PropsWithChildren<DateProps>): JSX.Element =
     questionnaire,
     answer,
     resources,
-    dispatch,
     path,
     language,
     promptLoginMessage,
@@ -60,7 +57,7 @@ const DateComponent = (props: React.PropsWithChildren<DateProps>): JSX.Element =
     onRenderMarkdown,
     children,
   } = props;
-
+  const dispatch = useDispatch<ThunkDispatch<GlobalState, void, NewValueAction>>();
   const getMaxDate = (): Date | undefined => {
     const maxDate = getExtension(Extensions.DATE_MAX_VALUE_URL, item);
     if (maxDate && maxDate.valueString) {
@@ -128,7 +125,6 @@ const DateComponent = (props: React.PropsWithChildren<DateProps>): JSX.Element =
 
   const labelText = getLabelText(item, onRenderMarkdown, questionnaire, resources);
   const subLabelText = getSublabelText(item, onRenderMarkdown, questionnaire, resources);
-  // console.log(maxDate);
 
   const itemControls = getItemControlExtensionValue(item);
   let element: JSX.Element | undefined = undefined;
@@ -187,16 +183,4 @@ const DateComponent = (props: React.PropsWithChildren<DateProps>): JSX.Element =
 };
 const withFormProps = ReactHookFormHoc(DateComponent);
 const withCommonFunctionsComponent = withCommonFunctions(withFormProps);
-const connectedComponent = connect(mapStateToProps, mapDispatchToProps, mergeProps)(withCommonFunctionsComponent);
-export default connectedComponent;
-
-// const shouldComponentUpdate = (nextProps: DateProps): boolean => {
-//   const responseItemHasChanged = responseItem !== nextProps.responseItem;
-//   const helpItemHasChanged = isHelpOpen !== nextProps.isHelpOpen;
-//   const answerHasChanged = answer !== nextProps.answer;
-//   const resourcesHasChanged = JSON.stringify(resources) !== JSON.stringify(nextProps.resources);
-//   const repeats = item.repeats ?? false;
-//   const error = error !== nextProps.error;
-
-//   return responseItemHasChanged || helpItemHasChanged || resourcesHasChanged || repeats || answerHasChanged || error;
-// };
+export default withCommonFunctionsComponent;