+++ src/components/referoLabel/ReferoLabel.tsx
@@ -4,8 +4,8 @@ import Label, { LabelText } from '@helsenorge/designsystem-react/components/Labe
 
 import SafeText from './SafeText';
 import SubLabel from './sublabel';
-import { getLabelText, getSublabelText } from '../../util';
-import { Resources } from '../../util/resources';
+import { getLabelText, getSublabelText } from '@/util';
+import { Resources } from '@/util/resources';
 
 type Props = {
   item: QuestionnaireItem;