+++ src/components/with-common-functions.tsx
@@ -40,6 +40,9 @@ import {
 } from '../util/refero-core';
 import { RenderContext } from '../util/renderContext';
 import { Resources } from '../util/resources';
+import { getAnswerIfDataReceiver, isEnableWhenEnabled } from '@/util/map-props';
+import { useSelector } from 'react-redux';
+import { getFormData } from '@/reducers/form';
 
 export interface WithCommonFunctionsProps {
   idWithLinkIdAndItemIndex: string;
@@ -84,7 +87,6 @@ export interface WithCommonFunctionsProps {
   ) => JSX.Element;
   onAnswerChange?: (newState: GlobalState, path: Array<Path>, item: QuestionnaireItem, answer: QuestionnaireResponseItemAnswer) => void;
   renderContext: RenderContext;
-  isHelpOpen?: boolean;
   onRenderMarkdown?: (item: QuestionnaireItem, markup: string) => string;
   fetchValueSet?: (
     searchString: string,
@@ -116,30 +118,32 @@ export default function withCommonFunctions<T extends WithCommonFunctionsProps>(
   WrappedComponent: React.ComponentType<T>
 ): React.ComponentType<T & EnhancedProps> {
   const WithCommonFunctions = (props: T): JSX.Element | null => {
+    const formData = useSelector((state: GlobalState) => getFormData(state));
+    const newAnswer = getAnswerIfDataReceiver(formData, props.item);
+
     const [isHelpVisible, setIsHelpVisible] = React.useState(false);
 
     const renderDeleteButton = (className?: string): JSX.Element | null => {
       if (!props.visibleDeleteButton) {
         return null;
       }
-
+      const hasAnwer = (answer: QuestionnaireResponseItemAnswer | QuestionnaireResponseItemAnswer[] | undefined): boolean => {
+        return !!answer && Object.keys(answer).length > 0;
+      };
       let mustShowConfirm: boolean = hasAnwer(props.answer);
 
       if (!mustShowConfirm && props.responseItem && props.responseItem.item) {
         mustShowConfirm = props.responseItem.item.some(item => (item ? hasAnwer(item.answer) : false));
       }
-      const idWithLinkIdAndItemIndex = props.item.linkId;
       return (
         <div className="page_refero__deletebutton-wrapper">
           <DeleteButton
-            idWithLinkIdAndItemIndex={idWithLinkIdAndItemIndex}
             className={className}
             item={props.item}
             path={props.path}
             resources={props.resources}
             mustShowConfirm={mustShowConfirm}
             onAnswerChange={props.onAnswerChange}
-            renderContext={props.renderContext}
           />
         </div>
       );
@@ -162,18 +166,12 @@ export default function withCommonFunctions<T extends WithCommonFunctionsProps>(
             key={`item_${item.linkId}_add_repeat_item`}
             resources={props.resources}
             item={item}
-            idWithLinkIdAndItemIndex={`${item.linkId}${createIdSuffix(path, index, item.repeats)}`}
             responseItems={response}
             parentPath={path}
-            renderContext={props.renderContext}
             disabled={item.type !== itemType.GROUP && !responseItem?.answer}
           />
         </div>
       );
-    };
-
-    const hasAnwer = (answer: QuestionnaireResponseItemAnswer | QuestionnaireResponseItemAnswer[] | undefined): boolean => {
-      return !!answer && Object.keys(answer).length > 0;
     };
 
     const renderHelpButton = (): JSX.Element | undefined => {
@@ -260,6 +258,7 @@ export default function withCommonFunctions<T extends WithCommonFunctionsProps>(
           response = getItemWithIdFromResponseItemArray(linkId, childAnswer[0].item);
         }
       }
+      const enable = !item || !item.enableWhen ? true : isEnableWhenEnabled(item.enableWhen, item.enableBehavior, path || [], formData);
       const renderedItems: Array<JSX.Element> = [];
       if (response && response.length > 0) {
         response.forEach((responseItem, index) => {
@@ -269,6 +268,7 @@ export default function withCommonFunctions<T extends WithCommonFunctionsProps>(
             <Comp
               idWithLinkIdAndItemIndex={idWithLinkIdAndItemIndex}
               key={'item_' + responseItem.linkId + createIdSuffix(path, index, item.repeats)}
+              enable={enable}
               pdf={pdf}
               language={props.language}
               includeSkipLink={props.includeSkipLink}
@@ -277,7 +277,7 @@ export default function withCommonFunctions<T extends WithCommonFunctionsProps>(
               item={item}
               questionnaire={props.questionnaire}
               responseItem={responseItem}
-              answer={getAnswerFromResponseItem(responseItem)}
+              answer={newAnswer ?? getAnswerFromResponseItem(responseItem)}
               resources={resources}
               containedResources={containedResources}
               path={createPathForItem(path, item, responseItem, index)}
@@ -319,10 +319,16 @@ export default function withCommonFunctions<T extends WithCommonFunctionsProps>(
 
       const renderedItems: Array<JSX.Element> = [];
       item.item.forEach(i => renderedItems.push(...renderItem(i, renderContext)));
+
       return renderedItems;
     };
 
-    if (!props.enable) {
+    const enable =
+      !props.item || !props.item.enableWhen
+        ? true
+        : isEnableWhenEnabled(props.item.enableWhen, props.item.enableBehavior, props.path || [], formData);
+
+    if (!enable) {
       return null;
     } else {
       return (
@@ -332,7 +338,7 @@ export default function withCommonFunctions<T extends WithCommonFunctionsProps>(
           renderRepeatButton={renderRepeatButton}
           renderHelpButton={renderHelpButton}
           renderHelpElement={renderHelpElement}
-          isHelpOpen={isHelpVisible}
+          enable={enable}
           {...props}
         >
           {renderChildrenItems(props.renderContext)}