+++ CHANGES.md
@@ -56,6 +56,19 @@
 
 - Rework of refero with new components
 - removed old components and added new from @helsenorge designsystem
+## 15.0.9
+
+---
+
+- Bugfix:
+  - fix a bug with sorting strings that looks like numbers. now trying to parse all stings as numbers and if they are numbers, sort as numbers
+
+## 15.0.8
+
+---
+
+- Bugfix:
+  - fix a bug in the table-hn2. Not being able to sort on the first column in the table
 
 ## 15.0.7